レビュー指摘の修正（バグ・表示ズレ6件、重複・効率4件）
バグ・表示ズレ:
- 同じ手数の別棋譜へ差し替えたとき再生位置・範囲選択が残る問題を修正
  （リセット判定を手数から棋譜配列の同一性に変更）
- 範囲切り出しノードの分岐元表示を「第n手から分岐」→「第s手〜第e手を切り出し」に
  変更（盤面は終点の局面のため表示と食い違っていた）
- 評価値を scheduleSave 経由の保存に変更し、blur 前の離脱でも
  flush 安全網（タブ切替・ページ離脱）で保存されるように修正
- 評価値の入力仕様を「整数部4桁＋小数第1位まで」に変更（ソフトにより
  評価値の形式が異なるため）。DBカラムを numeric(5,1) に変更
- 棋譜の名前変更が保存完了を待ってから閉じるように修正
- 1手も読めない棋譜の「第0手までを読み込みました」表記を修正し、
  0手棋譜は保存対象外に

重複・効率:
- fetchKifu の select を絞り、未使用の source_text（KIF原文）を転送しない
- 削除確認モーダルを ConfirmDeleteModal として uiParts へ共通化
  （ツリー削除・棋譜削除で共用）
- 棋譜プレビュー盤面を KifuPreviewBoard として uiParts へ共通化
  （ライブラリのプレビュー・取り込みモーダルで共用）
- 分岐ノード作成処理を createBranchNode に集約
  （「この局面で分岐」「範囲切り出し」で共用）

+++ supabase/migrations/20260718_add_turn_evaluation.sql
@@ -2,6 +2,7 @@
 -- 盤面の手番（先手番/後手番）と評価値をノードに追加
 --   turn:       'sente' | 'gote' | null（未設定）。手動選択で保存する
 --   evaluation: 符号込みの評価値（例: 300 = 先手300点良し / -300 = 後手300点良し）
+--               ソフトにより形式が異なるため小数第1位まで対応（整数部4桁 = ±9999.9）
 --   あわせて copy_tree RPC を再作成し、INSERT列リストに新カラムを追加する
 --   （copy_tree はカラム明示型のため、nodes へのカラム追加時は必ず追従が必要）
 -- Supabase の SQL Editor で実行してください
@@ -11,7 +12,7 @@ alter table public.nodes
   add column if not exists turn text;
 
 alter table public.nodes
-  add column if not exists evaluation int;
+  add column if not exists evaluation numeric(5,1);
 
 create or replace function public.copy_tree(p_source_tree_id uuid, p_new_name text default null)
 returns uuid
